window: Use sinf instead of sin

+++ libtuna/window.c
@@ -38,6 +38,6 @@ void window_init_sine(float * window, uint length)
 	assert(window);
 
 	for (i = 0; i < length; i++) {
-		window[i] = scale * sin(M_PI * i / length);
+		window[i] = scale * sinf(M_PI * i / length);
 	}
 }
